+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-# 0.9.0 (not yet)
+# [0.1.1]  - 2022-06-27
+
+### Fixed
+
+- Lib is now a package to use and not a program
+- TCPServer Session returns RemoteAddr
+
+# [0.1.0]  - 2022-06-27
 
 First release with basic features
 
-* TCP Client and Server
-* Transmission and Sending without a protocol
+### Added
+
+- TCP Client and Server
+- Transmission and Sending without a protocol